requirements.txt
cloudant==2.0.0b1
kafka-python
